Changed outfile to HTML5, fixed array printing bug
In FieldInfo, there was a bug in the way arrays were
parsed from the bytes in the field. The for loop only checked
the loop counter against the length of the array of items,
not against (Length * Size) as was correct. This led to,
usually, only printing the first element (and the rest being 0s),
or in other cases, printing values in the wrong places, printing
wrong values, and/or skipping values.