Major restructure of core lib, added browser
Restructured the core library and removed the flattened fields in
PopulatedStructure. Changed FieldInfo to Field<T>. Created interfaces IField,
IField<T>, and IStructField, which are implemented appropriately by
PopulatedStructure and Field<T>. Removed GetValueAs* functions from
Field<T> in lieu of both generic and dynamic Value properties. Embedded
structures now use PopulatedStructure instead of a series of flattened
Field<T> objects. Added a Dereference function to Field as a shortcut to
"dereferencing" the pointer yourself. Updated HTMLRendererDriver and
XNADriver appropriately.

Added a GUI browser in BlenderFileBrowser, which allows you to browse
through the file in a TreeView. Not really quite finished; it's a little
unintuitive and messy.