Renamed types in corelib
Types renamed:
  BlenderField -> FieldDefinition
  PopulatedStructure -> Structure
  SDNAStructure -> StructureDefinition
  BlenderType -> TypeDefinition

All non-pipeline projects have been updated to reflect this change. Pipeline
projects have been unloaded for this commit, as they are currently horribly broken.

The purpose of this is to better define the seperation between a
definition (or prototype) of something in SDNA and its implementation, as
well as change naming conventions to make the prototypes and
implementations more consistant. The current convention is to define
prototypes as ending with "Definition" and to define implementations (if
one exists) as omitting the "Definition" suffix.