modified corelib and main projects
They don't work, the CLR will not try an implicit cast to a type when
direct invocation fails. So, if a Field holds a ulong, and supports
conversion to a ulong, but is called as methodThatExpectsUlong(field), it
will fail because field isn't a ulong (even though it can be converted to
one). Working on a fix.